adding get_path to schema validatator
Also moved get_container
